allow addition of multiple edocs:
This is a little slow because it needs to re-render the whole page
after reach edoc. This could be solved in two ways:
  * use a formset to display the existing edocs like w/ notes/tags
  * make this page load fast enough that it doesnt matter
