+++ escalate/core/forms/forms.py
@@ -17,6 +17,9 @@ if version.parse(django.__version__) < version.parse('3.1'):
 else:
     from django.forms import JSONField
 
+from crispy_forms.helper import FormHelper
+from crispy_forms.layout import Layout, Submit, Row, Column, Hidden, Field
+
 dropdown_attrs = {'class': 'selectpicker',
                   'data-style': 'btn-outline-primary', 'data-live-search': 'true'}
 
@@ -632,3 +635,19 @@ class VesselForm(forms.ModelForm):
                 'placeholder': 'Well number'
             }),
         }
+
+
+class UploadFileForm(forms.Form):
+    file = forms.FileField(label='Upload file', required=False)
+    #upload_edoc = forms.Select(label='Submit')
+    @staticmethod
+    def get_helper():
+        helper = FormHelper()
+        helper.form_class = 'form-horizontal'
+        helper.label_class = 'col-lg-2'
+        helper.field_class = 'col-lg-8'
+        helper.layout = Layout(
+            Row(Column(Field('file'))),
+            #Row(Column(Submit('upload_edoc', 'Submit'))),
+        )
+        return helper
